// src/screens/logger/report/type.d.ts
export interface columnInterface {
  createdAt: string;
  id: string;
  isPermanentDisable: boolean;
  isSortBtn: boolean;
  name: string;
  originalName: string;
  positionName: string;
  position: number;
  updatedAt: string;
  referancKeyName?: string;
  currentName: string;
  isActive: string | boolean;
}